% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cities}
\alias{cities}
\title{Cities dataset
   Overview of cities included in the shiny app, their respective countries and
   city names in english. This data frame was created manually and is mainly used
   for internal data wrangling.}
\format{
## `cities`
A data frame with 82 rows and 3 columns:
\describe{
  \item{country}{Country name}
  \item{city}{City name}
  \item{city_english}{City name in english as shown in the shiny user interface}
}
}
\usage{
cities
}
\description{
Cities dataset
   Overview of cities included in the shiny app, their respective countries and
   city names in english. This data frame was created manually and is mainly used
   for internal data wrangling.
}
\keyword{datasets}
