% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{city_prices}
\alias{city_prices}
\title{Living expenses & product prices of selected European countries}
\format{
A data frame with 1186 rows and 7 columns:
\describe{
  \item{country}{Country name}
  \item{city}{City name}
  \item{item_name}{Name of product}
  \item{category}{Category which the product belongs to. Values are "Housing", "Groceries", "Alcoholic Beverages", "Leisure time and going out", "Transportation"}
  \item{min}{lowest price of product in city; in EURO}
  \item{avg}{highest price of product in city; in EURO}
  \item{max}{average price of product in city; in EURO}
}
}
\source{
Cost of Living & Prices API: <https://cost-of-living-and-prices.p.rapidapi.com/prices>
}
\usage{
city_prices
}
\description{
This data frame contains prices of various goods in European cities.
Prices are reported in Euro and were retrieved in May 2023.
Data was taken from Cost of Living & Prices API. Copyright therefore
does not fall under the MIT license but belongs to Cost of Living & Prices API.
}
\keyword{datasets}
