% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_living_expenses.R
\name{filter_prices}
\alias{filter_prices}
\title{Get living expenses of two European cities}
\usage{
filter_prices(city1, city2)
}
\arguments{
\item{city1}{First city to be retrieved.

   Valid city names are:
   Aalborg, Amsterdam, Antwerp, Athens, Barcelona, Belfast, Berlin, Białystok,
   Bologna, Bordeaux, Braga, Bratislava, Brussels, Bucharest, Budapest,
   Burgas, Cardiff, Copenhagen, Dortmund, Dublin, Essen, Gdańsk, Geneva,
   Glasgow, Graz, Groningen, Hamburg, Helsinki, Heraklion, Košice, Leipzig,
   Liège, Lille, Lisbon, Ljubljana, London, Madrid, Málaga, Malmö,
   Manchester, Marseille, Miskolc, Munich, Naples, Nicosia, Oslo, Ostrava,
   Oulu, Oviedo, Palermo, Paris, Podgorica, Prague, Rennes, Reykjavík,
   Riga, Rome, Rostock, Rotterdam, Skopje, Sofia, Stockholm, Strasbourg,
   Tallinn, Tirana, Turin, Valletta, Verona, Vienna, Vilnius, Warsaw,
   Zagreb, Zurich}

\item{city2}{Second city to be retrieved}
}
\value{
A data frame containing city name, price category, product name, and
price in euros
}
\description{
This function returns a data frame containing the prices of
various goods in 2 European cities of your choice. Prices are provided in €
and were taken from the Cost of living and prices:
https://rapidapi.com/traveltables/api/cost-of-living-and-prices/#
}
