% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_quality_of_life.R
\name{filter_QoL_comparison}
\alias{filter_QoL_comparison}
\title{Function to filter quality of life rating of 2 chosen cities on one
chosen variable}
\usage{
filter_QoL_comparison(country1, country2, city1, city2, criterion)
}
\arguments{
\item{country1}{Country of first city of choice

   \emph{Valid country names are}:
   Albania, Austria, Belgium, Bulgaria, Croatia, Cyprus, Czechia, Denmark,
   England, Estonia, Finland, France, Germany, Greece, Hungary, Iceland,
   Ireland, Italy, Latvia, Lithuania, Luxembourg, Malta, Montenegro,
   Netherlands, North Macedonia, Northern Ireland, Norway, Poland, Portugal,
   Romania, Scotland, Serbia, Slovakia, Slovenia, Spain, Sweden, Switzerland,
   Turkey, Wales}

\item{country2}{Country of second city of choice}

\item{city1}{First city of choice.

   \emph{Valid city names are}:
   Aalborg, Amsterdam, Antwerp, Athens, Barcelona, Belfast, Berlin, Białystok,
   Bologna, Bordeaux, Braga, Bratislava, Brussels, Bucharest, Budapest,
   Burgas, Cardiff, Copenhagen, Dortmund, Dublin, Essen, Gdańsk, Geneva,
   Glasgow, Graz, Groningen, Hamburg, Helsinki, Heraklion, Košice, Leipzig,
   Liège, Lille, Lisbon, Ljubljana, London, Madrid, Málaga, Malmö, Manchester,
   Marseille, Miskolc, Munich, Naples, Nicosia, Oslo, Ostrava, Oulu, Oviedo,
   Palermo, Paris, Podgorica, Prague, Rennes, Reykjavík, Riga, Rome, Rostock,
   Rotterdam, Skopje, Sofia, Stockholm, Strasbourg, Tallinn, Tirana,
   Turin, Valletta, Verona, Vienna, Vilnius, Warsaw, Zagreb, Zurich}

\item{city2}{Second city}

\item{criterion}{Criterion to return; i.e, category that inhabitants were asked
   to rate. The following criteria were assessed:

   `LGBTQI` (LGBTQI+ friendliness),
   `air` (Air quality),
   `clean` (Cleanness),
   `culture` (Diverse cultural offer),
   `edu` (Education system),
   `greenery` (Green spaces),
   `health` (Health system),
   `noise` (Noise level),
   `racial` (Quality of life for ethnic minorities),
   `safety` (Safety),
   `sport` (Sport facilities),
   `transp` (Public transport),
   `satisfaction` (general satisfaction with city)}
}
\value{
A data frame
}
\description{
This function returns a data frame containing the satisfaction of
   inhabitants of 2 European cities of your choice with a certain criterion of
   quality of life.
}
\examples{
filter_QoL_comparison("Germany", "Netherlands", "Berlin", "Amsterdam", "LGBTQI")

}
