% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{life_satisfaction}
\alias{life_satisfaction}
\title{World Happiness Report 2023 data for selected European countries}
\format{
A data frame with 1162 rows and 4 columns:
\describe{
  \item{Entity}{Country name}
  \item{Year}{Year of measurement}
  \item{Life_satisfaction}{Country mean of self-reported life satisfaction on a scale of 0 to 10}
  \item{Code}{Country Code (used for plotting map in shiny)}
}
}
\source{
Helliwell, J. F., Layard, R., Sachs, J. D., De Neve, J.-E., Aknin, L. B., & Wang, S. (Eds.). (2023).
    World Happiness Report 2023., <https://ourworldindata.org/grapher/happiness-cantril-ladder>
}
\usage{
life_satisfaction
}
\description{
This data frame contains a subset of the data of the World Happiness Report 2023.
Data were collected yearly from 2003 to 2021 and contain country-wide means
of self-reported life satisfaction on a scale of 0 to 10.
}
\keyword{datasets}
