% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_life_satisfaction.R
\name{plot_lifeSat}
\alias{plot_lifeSat}
\title{Plot the trajectories of self-reported life satisfaction of two chosen European
countries between 2003 and 2021}
\usage{
plot_lifeSat(country1, country2)
}
\arguments{
\item{country1}{First country to be plotted.

   Valid country names are:

   Albania, Austria, Belgium, Bulgaria, Croatia, Cyprus, Czechia,
   Denmark, England, Estonia, Finland, France, Germany, Greece,
   Hungary, Iceland, Ireland, Italy, Latvia, Lithuania, Luxembourg,
   Malta, Montenegro, Netherlands, North Macedonia, Northern Ireland,
   Norway, Poland, Portugal, Romania, Scotland, Serbia, Slovakia,
   Slovenia, Spain, Sweden, Switzerland, Turkey, Wales}

\item{country2}{Second country to be plotted}
}
\value{
A line plot
}
\description{
Data is taken form the World Happiness Report (Helliwell et al., 2003)
}
\examples{
plot_lifeSat("Netherlands", "Turkey")

}
