% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qualityOL}
\alias{qualityOL}
\title{Report on the Quality of life in European Cities data}
\format{
A data frame with 1162 rows and 4 columns:
\describe{
  \item{country}{Country name}
  \item{city}{City name}
  \item{variable}{Criterion participants were asked to rate}
  \item{percentage}{percentage of participants reporting to be 'rather satisfied' or 'very satisfied' with the respective criterion in their city}
}
}
\source{
<https://ec.europa.eu/regional_policy/information-sources/maps/quality-of-life_en>
}
\usage{
qualityOL
}
\description{
Results of the Report on the Quality of life in European Cities, reported
in 2019 and published in 2020.
}
\keyword{datasets}
