% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shall_I_move.R
\name{shall_I_move}
\alias{shall_I_move}
\title{Start the 'shall I move' shiny app}
\usage{
shall_I_move()
}
\arguments{
\item{NONE}{}
}
\value{
A shiny app
}
\description{
This function starts a shiny app which allows for descriptive
comparison of european cities regarding their quality of life, life satisfaction
of inhabitants, and living expenses.
}
\examples{
shall_I_move()

}
