% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_living_expenses.R
\name{filter_prices}
\alias{filter_prices}
\title{Get living expenses of two European cities}
\usage{
filter_prices(city1, city2)
}
\arguments{
\item{city1}{First city to be retrieved.

   Valid city names are: "Barcelona", "Athens", "Amsterdam", "Bologna" "Bordeaux", "Brussels", "Helsinki"
   "Copenhagen", "Lisbon", "Dublin", "Hamburg", "Groningen", "Berlin", "Aalborg",
   "Bratislava", "Bucharest", "Budapest", "Madrid", "Rome", "Oslo", "Paris",
   "Prague", "Stockholm", "Sofia", "Rotterdam", "Tallinn", "Vilnius", "Zagreb",
   "Warsaw", "Verona", "Zurich", "Vienna", "Ljubljana", "Lille", "Malmö",
   "Munich", "Riga", "Reykjavík", "Valletta", "Gdańsk", "Oulu", "Leipzig",
   "Braga", "Marseille", "Antwerp", "Tirana", "Strasbourg", "Skopje", "Podgorica",
   "Białystok", "Nicosia", "Miskolc", "Málaga", "Košice", "Graz", "Heraklion",
   "Oviedo", "Geneva", "Burgas", "Ostrava", "Turin", "Rostock", "Rennes",
   "Palermo", "Naples", "Manchester", "London", "Glasgow", "Cardiff", "Belfast",
   "Dortmund", "Essen", "Liège"}

\item{city2}{Second city to be retrieved}
}
\value{
A data frame containing city name, price category, product name, and
price in euros
}
\description{
This function returns a data frame containing the prices of
various goods in 2 European cities of your choice. Prices are provided in €
and were taken from the Cost of living and prices:
https://rapidapi.com/traveltables/api/cost-of-living-and-prices/#
}
